extract sort function
